+++ assignment-3/src/components/AddBookModal/AddBookModal.types.ts
@@ -1,10 +1,9 @@
-import { FormInstance } from 'antd';
-import { Book } from '../../pages/Homepage/Homepage.types';
+import { FormInstance } from 'antd'
+import { Book } from '../../pages/Homepage/Homepage.types'
 
 export interface AddBookModalProps {
-  form: FormInstance<any>;
-  openModal: boolean;
-  handleAddBook: (value: Book) => void;
-  handleCloseModal: () => void;
-  topics: String[];
+  form: FormInstance<any>
+  openModal: boolean
+  handleAddBook: (value: Book) => void
+  handleCloseModal: () => void
 }
